Finished add_xml and update_xml specs
