+++ lib/qbwc/request/products.rb
@@ -88,7 +88,7 @@ module QBWC
           <<~XML
             <ItemInventoryAddRq requestID="#{session_id}">
                <ItemInventoryAdd>
-                #{product_xml(product, params, config, false)}
+                #{product_xml(product, config, false)}
                </ItemInventoryAdd>
             </ItemInventoryAddRq>
           XML
@@ -100,7 +100,7 @@ module QBWC
                 <ItemInventoryMod>
                   <ListID>#{product['list_id']}</ListID>
                   <EditSequence>#{product['edit_sequence']}</EditSequence>
-                  #{product.key?('active') ? product_only_touch_xml(product, params) : product_xml(product, params, config, true)}
+                  #{product.key?('active') ? product_only_touch_xml(product, params) : product_xml(product, config, true)}
                 </ItemInventoryMod>
             </ItemInventoryModRq>
           XML
@@ -113,7 +113,7 @@ module QBWC
           XML
         end
 
-        def product_xml(product, params, config, is_mod)
+        def product_xml(product, config, is_mod)
           <<~XML
             <Name>#{product_identifier(product)}</Name>
             #{add_barcode(product)}
@@ -216,10 +216,11 @@ module QBWC
         end
 
         def add_fields(object, mapping, config, is_mod)
+          object = object.with_indifferent_access
           fields = ""
           mapping.each do |map_item|
-            return "" if object[:mod_only] && object[:mod_only] != is_mod
-            return "" if object[:add_only] && object[:add_only] == is_mod
+            next if map_item[:mod_only] && map_item[:mod_only] != is_mod
+            next if map_item[:add_only] && map_item[:add_only] == is_mod
 
             if map_item[:is_ref]
               fields += add_ref_xml(object, map_item, config)
@@ -251,8 +252,8 @@ module QBWC
             return "<#{qbe_field_name}><ListID>#{flowlink_field['list_id']}</ListID></#{qbe_field_name}>"
           end
           full_name = flowlink_field ||
-                                config[mapping[:flowlink_name]] ||
-                                config["quickbooks_#{mapping[:flowlink_name]}"]
+                                config[mapping[:flowlink_name].to_sym] ||
+                                config["quickbooks_#{mapping[:flowlink_name]}".to_sym]
 
           full_name.nil? ? "" : "<#{qbe_field_name}><FullName>#{full_name}</FullName></#{qbe_field_name}>"
         end
