+++ spec/qbwc/request/product_fixtures/invproduct_from_flowlink.json
@@ -1,0 +1,35 @@
+{
+  "id": "FE1221",
+  "sku": "FE1221",
+  "product_id": "FE1221",
+  "is_active": true,
+  "list_id": "FE1221",
+  "class_name": "my class",
+  "parent_name": "my parent",
+  "manufacturer_part_number": "man_part-num",
+  "unit_of_measure": "feet",
+  "is_tax_included": false,
+  "sales_tax_code_name": "tax",
+  "description": "Crazy product description",
+  "price": 155.0,
+  "inventory_account": "Inventory Asset",
+  "income_account": "Income Acct",
+  "purchase_description": "Crazy purchase description",
+  "force_uom_change": true,
+  "cost": 62.0,
+  "apply_cogs_account_ref_to_existing_txns": true,
+  "apply_income_account_ref_to_existing_txns": true,
+  "purchase_tax_code_name": "purchase tax",
+  "preferred_vendor_name": {
+    "list_id": "1029203902"
+  },
+  "reorder_point": 99,
+  "max": 10000,
+  "quantity": 1000,
+  "total_value": 20000,
+  "inventory_date": "2019-09-24T17:24:16-05:00",
+  "barcode_value": "3990239-399239",
+  "assign_barcode_even_if_used": true,
+  "external_guid": "0120i-f23f23-f2im39-199f33m993mf",
+  "edit_sequence": "19209j3od-d9292"
+}
